+++ Package.swift
@@ -15,17 +15,11 @@ let package = Package(
             name: "SwiftRest",
             targets: ["SwiftRest"]),
     ],
-    dependencies: [
-        .package(url: "https://github.com/ricky-stone/SwiftKey.git", from: "1.1.0")
-    ],
     targets: [
         // Targets are the basic building blocks of a package, defining a module or a test suite.
         // Targets can depend on other targets in this package and products from dependencies.
         .target(
-            name: "SwiftRest",
-            dependencies: [
-                .product(name: "SwiftKey", package: "SwiftKey")
-            ]),
+            name: "SwiftRest"),
         .testTarget(
             name: "SwiftRestTests",
             dependencies: ["SwiftRest"]